// src/types/olo-api/olo-api.enums.ts
export enum DeliveryModeEnum {
  delivery = 'delivery',
  dispatch = 'dispatch',
  curbside = 'curbside',
  pickup = 'pickup',
  dinein = 'dinein',
  drivethru = 'drivethru'
}